fix: Prepare for StackSet but not use it at this point, don't set IAM Password Policy
It doesn't seem to work out of the box to push an updated iam password policy through a stackset and requires more lambda logic to achieve it. So I'm dropping this for now, I don't plan to continue to use IAM User either way. Creation of them is disabled too. So, it was just "in case of" icing on the cake to have it.

Refs: https://github.com/aws-samples/iam-password-policy-deployment-via-cloudformation-stacksets